kubectl-datadog: extract EKS helpers to common/eks/
Continue the redistribution of `guess/` started by PRs #2980 and #3000:
move every helper that operates on an EKS cluster object or the EKS API
into a dedicated `common/eks/` package. After this commit, `common/eks/`
contains:
  - authmode.go (was guess/clusterauthmode.go)
  - podidentityagent.go (was guess/ekspodidentityagent.go)
  - oidcprovider.go (was guess/oidcprovider.go)
  - privatesubnets.go (was guess/privatesubnets.go)

Import alias `commoneks` is used to avoid colliding with the existing
`github.com/aws/aws-sdk-go-v2/service/eks` (aliased `eks`).
