+++ cmd/kubectl-datadog/autoscaling/cluster/common/eks/oidcprovider_test.go
@@ -1,4 +1,4 @@
-package guess
+package eks
 
 import (
 	"context"
