+++ cmd/kubectl-datadog/autoscaling/cluster/common/eks/podidentityagent.go
@@ -1,4 +1,4 @@
-package guess
+package eks
 
 import (
 	"context"
